check for not adding duplicates

+++ src/js/favorites.ts
@@ -11,14 +11,25 @@ export let favorites: IBook[] = JSON.parse(
 
 export function addToFavorites(book: IBook) {
   // kanske plocka ut 3-4 egenskaper och göra favorites till datatyp: IFavoriteBooks []
-  favorites.push(book);
-  localStorage.setItem("favorites", JSON.stringify(favorites));
+  const bookExistsInFavorites = favorites.some((fav) => fav.id === book.id);
+  console.log(bookExistsInFavorites);
+
+  if (!bookExistsInFavorites) {
+    favorites.push(book);
+    localStorage.setItem("favorites", JSON.stringify(favorites));
+  }
 }
 
 favPopoverBtn?.addEventListener("click", updateFavoritesUI);
 
 function updateFavoritesUI() {
-  favoritesPopover.innerHTML = favorites
-    .map((fav) => `<div>${fav.title}</div>`)
-    .join("");
+  favoritesPopover.innerHTML = "";
+
+  favorites.forEach((fav) => {
+    const title = document.createElement("h3");
+    title.textContent = fav.title;
+    const removeBtn = document.createElement("button");
+    removeBtn.textContent = "Remove";
+    favoritesPopover.append(title, removeBtn);
+  });
 }
